1st pass - remove ContractUpgrade public API calls from VaultService into self contained ContractUpgradeService.
