+++ core/src/main/kotlin/net/corda/core/node/services/VaultService.kt
@@ -180,24 +180,6 @@ interface VaultService {
         return updates.filter { it.consumed.any { it.ref == ref } }.toFuture()
     }
 
-    /** Get contracts we would be willing to upgrade the suggested contract to. */
-    // TODO: We need a better place to put business logic functions
-    fun getAuthorisedContractUpgrade(ref: StateRef): Class<out UpgradedContract<*, *>>?
-
-    /**
-     * Authorise a contract state upgrade.
-     * This will store the upgrade authorisation in the vault, and will be queried by [ContractUpgradeFlow.Acceptor] during contract upgrade process.
-     * Invoking this method indicate the node is willing to upgrade the [state] using the [upgradedContractClass].
-     * This method will NOT initiate the upgrade process. To start the upgrade process, see [ContractUpgradeFlow.Instigator].
-     */
-    fun authoriseContractUpgrade(stateAndRef: StateAndRef<*>, upgradedContractClass: Class<out UpgradedContract<*, *>>)
-
-    /**
-     * Authorise a contract state upgrade.
-     * This will remove the upgrade authorisation from the vault.
-     */
-    fun deauthoriseContractUpgrade(stateAndRef: StateAndRef<*>)
-
     /**
      *  Add a note to an existing [LedgerTransaction] given by its unique [SecureHash] id
      *  Multiple notes may be attached to the same [LedgerTransaction].
