+++ node/src/main/kotlin/net/corda/node/services/upgrade/ContractUpgradeServiceImpl.kt
@@ -1,0 +1,26 @@
+package net.corda.node.services.upgrade
+
+import net.corda.core.contracts.StateAndRef
+import net.corda.core.contracts.StateRef
+import net.corda.core.contracts.UpgradedContract
+import net.corda.core.node.services.ContractUpgradeService
+
+class ContractUpgradeServiceImpl : ContractUpgradeService {
+
+    // TODO : Persist this in DB.
+    private val authorisedUpgrade = mutableMapOf<StateRef, Class<out UpgradedContract<*, *>>>()
+
+    override fun getAuthorisedContractUpgrade(ref: StateRef) = authorisedUpgrade[ref]
+
+    override fun authoriseContractUpgrade(stateAndRef: StateAndRef<*>, upgradedContractClass: Class<out UpgradedContract<*, *>>) {
+        val upgrade = upgradedContractClass.newInstance()
+        if (upgrade.legacyContract != stateAndRef.state.data.contract.javaClass) {
+            throw IllegalArgumentException("The contract state cannot be upgraded using provided UpgradedContract.")
+        }
+        authorisedUpgrade.put(stateAndRef.ref, upgradedContractClass)
+    }
+
+    override fun deauthoriseContractUpgrade(stateAndRef: StateAndRef<*>) {
+        authorisedUpgrade.remove(stateAndRef.ref)
+    }
+}
